last changes tested

+++ src/employee/EmployeeRole.java
@@ -4,14 +4,10 @@
  */
 package employee;
 
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.Scanner;
 
 /**
  *
@@ -29,19 +25,19 @@ public class EmployeeRole {
     public void addProduct(String productID, String productName, String manufacturerName, String supplierName,
             int quantity, float price) throws IOException {
         Product p = new Product(productID, productName, manufacturerName, supplierName, quantity, price);// composition
-        
-            productsDatabase.insertRecord(p);
-            productsDatabase.saveToFile();
-      
+
+        productsDatabase.insertRecord(p);
+        productsDatabase.saveToFile();
+
     }
 
     public Product[] getListOfProducts() {
         ArrayList<Product> p = new ArrayList<>();
         // composition, can not inherit from more than one class, so we used composition
         // here.
-      
-            productsDatabase.readFromFile();
-        
+
+        productsDatabase.readFromFile();
+
         p = productsDatabase.returnAllRecords();
         Product[] products = new Product[p.size()];
         p.toArray(products);
@@ -51,10 +47,10 @@ public class EmployeeRole {
 
     public CustomerProduct[] getListOfPurchasingOperations() {
         ArrayList<CustomerProduct> c = new ArrayList<>();
-   
-            customerProductDatabase.readFromFile();// composition, can not inherit from more than one class, so we used
-                                                   // composition here.
-      
+
+        customerProductDatabase.readFromFile();// composition, can not inherit from more than one class, so we used
+                                               // composition here.
+
         c = customerProductDatabase.returnAllRecords();
         CustomerProduct[] customer = new CustomerProduct[c.size()];
         c.toArray(customer);
@@ -64,74 +60,73 @@ public class EmployeeRole {
 
     public boolean purchaseProduct(String customerSSN, String productID, LocalDate purchaseDate) throws IOException {
         Product p = productsDatabase.getRecord(productID);
-        if(p!=null) 
-        {if (p.getQuantity() == 0) // checks if the product is in stock.
-        {
-            System.out.println("The product is out of stock.");
-            return false;
-        } else {
-            p.sellUnit();// decrements quantity by one.
-            CustomerProduct c = new CustomerProduct(customerSSN, productID, purchaseDate);
-           
+        if (p != null) {
+            if (p.getQuantity() == 0) // checks if the product is in stock.
+            {
+                System.out.println("The product is out of stock.");
+                return false;
+            } else {
+                p.sellUnit();// decrements quantity by one.
+                CustomerProduct c = new CustomerProduct(customerSSN, productID, purchaseDate);
+
                 customerProductDatabase.insertRecord(c);
                 customerProductDatabase.saveToFile();
                 productsDatabase.deleteRecord(p.getSearchKey());
                 productsDatabase.insertRecord(p);
                 productsDatabase.saveToFile();
-            
-            System.out.println("The purchase was made successfully!");
-            return true;
+
+                System.out.println("The purchase was made successfully!");
+                return true;
+            }
+
+        } else {
+            System.out.println("Product not found!");
+            return false;
+        }
+    }
+
+    public double returnProduct(String customerSSN, String productID, LocalDate purchaseDate, LocalDate returnDate)
+            throws IOException {
+        // checl if return date is before purchase or not
+
+        if (returnDate.isBefore(purchaseDate)) {
+            System.out.println("Return date is incorrect.");
+            return -1;
         }
 
-    } 
-        else{ System.out.println("Product not found!");
-        return false;}
-    }
-       
-
-    public double returnProduct(String customerSSN, String productID, LocalDate purchaseDate, LocalDate returnDate) throws IOException {
-        // checl if return date is before purchase or not
-        
-        if (returnDate.isBefore(purchaseDate))
-        {System.out.println("Return date is incorrect.");
-            return -1;}
-
-       CustomerProduct c=new CustomerProduct(customerSSN,productID,purchaseDate);
-           customerProductDatabase.readFromFile();// composition, can not inherit from more than one class, so we used
-                                                   // composition here.
-        if (customerProductDatabase.getRecord(c.getSearchKey())==null)
-        {System.out.println("Record not found!");
-            return -1;}
+        CustomerProduct c = new CustomerProduct(customerSSN, productID, purchaseDate);
+        customerProductDatabase.readFromFile();// composition, can not inherit from more than one class, so we used
+                                               // composition here.
+        if (customerProductDatabase.getRecord(c.getSearchKey()) == null) {
+            System.out.println("Record not found!");
+            return -1;
+        }
         // check if product is found or not
 
-        if (ChronoUnit.DAYS.between(purchaseDate, returnDate) > 14)
-        {System.out.println("Exceeded the 14 days.");
-            return -1;}
+        if (ChronoUnit.DAYS.between(purchaseDate, returnDate) > 14) {
+            System.out.println("Exceeded the 14 days.");
+            return -1;
+        }
         // ChronoUnit.DAYS it measure times in units like days,month, years
         // .between calculate the difference in time between two local objects
         productsDatabase.readFromFile();
 
         Product productMatch = productsDatabase.getRecord(productID);
 
-        if (productMatch != null)
-        {productMatch.returnUnit();
-        
-              
+        if (productMatch != null) {
+            productMatch.returnUnit();
+
             productsDatabase.deleteRecord(productID);
             productsDatabase.insertRecord(productMatch);
             productsDatabase.saveToFile();
             customerProductDatabase.deleteRecord(c.getSearchKey());
             customerProductDatabase.saveToFile();
             System.out.println("The product was returned successfully!");
-           return productMatch.getPrice();
-        }
-        else
-        return -1;
+            return productMatch.getPrice();
+        } else
+            return -1;
 
-
-        
         // update both customerproduct anf product databases files
-     
 
     }
 
@@ -145,25 +140,26 @@ public class EmployeeRole {
                     return true;
                 } // check if its paid or not
                 record.setPaid(true);
-               
-                    customerProductDatabase.deleteRecord(record.getSearchKey());// deletes old record
-                    customerProductDatabase.insertRecord(record);// insert the new updated record
-                    customerProductDatabase.saveToFile();
-                    System.out.println("Payment made successfully!");
-                
-                return true;
-            } // not paid then set it to paid
+
+                customerProductDatabase.deleteRecord(record.getSearchKey());// deletes old record
+                customerProductDatabase.insertRecord(record);// insert the new updated record
+                customerProductDatabase.saveToFile();
+                System.out.println("Payment made successfully!");
+
+                return true; // not paid then set it to paid
+            }
         }
-        return false;// record is not found
+        System.out.println("Please check ssn and purchase date");// record is not found
+        return false;
     }
 
     public void logout() {
-       try{
+        try {
             productsDatabase.saveToFile();
-           customerProductDatabase.saveToFile();}
-       catch(IOException e)
-       {System.out.println(e.getMessage());}
-        
+            customerProductDatabase.saveToFile();
+        } catch (IOException e) {
+            System.out.println(e.getMessage());
+        }
+
     }
 }
-
